Add package, more OOP(except parser)
