+++ data/tasks.txt
@@ -1,4 +1,2 @@
-T | 1 | vvvv
-D | 0 | sadf  | sa
-E | 1 | sdf  | 234
-T | 0 | chuash ihong
+D | 1 | a  | 2022-01-12
+T | 0 | 1
