+++ src/main/java/duke/Deadline.java
@@ -1,3 +1,4 @@
+package duke;
 public class Deadline extends Task {
 
     protected String by;
