+++ src/main/java/duke/DeleteException.java
@@ -1,3 +1,4 @@
+package duke;
 public class DeleteException extends DukeException{
     public DeleteException(String errorMessage) {
         super(errorMessage);
